gdpr_checker_service/requirements.txt
fastapi
uvicorn[standard]
openai
langchain
chromadb
python-dotenv
langchain-community
